Update _157-Unset The K-th bit of a given number
Given a number N and a value K. From the right, unset/clear the Kth bit in the binary representation of N.If K-th bit is 1, then clear it to 0 and if it is 0 then leave it unchanged.

The position of Least Significant Bit(or last bit) is 1, the second last bit is 2 and so on. 

Examples: 
 

Input: N = 5, K = 1
Output: 4
5 is represented as 101 in binary
and has its first bit 1, so clearing
it will result in 100 i.e. 4.

Input: N = 5, K = 2
Output: 5
5 is represented as 101 in binary
and has its second bit is already 0,
so clearing it will result in 101 i.e. 5.


Expected Time Complexity: O(1)
Expected Auxiliary Space: O(1)

Constraints:
1 <= N <= 10^9
1 <= K < X, where X is the number of bits in the binary representation of N.