Update _158-Toggling K-th Bit Of A Number
For a given number n, if K-th bit is 0, then toggle it to 1 and if it is 1 then, toggle it to 0. Please note that k = 1 means the rightmost bit.

Examples : 
 
Input : n = 5, k = 1
Output : 4
5 is represented as 101 in binary
and has its first bit 1, so toggling 
it will result in 100 i.e. 4.

Input : n = 2, k = 3
Output : 6

Input : n = 75, k = 4
Output : 67