Create _204-Calculate Power using Recursion
1. You are given a number x.
2. You are given another number n.
3. You are required to calculate x raised to the power n. Don't change the signature of power function .

Note -> The online judge can't force you to write the function recursively but that is what the spirit of question is. Write recursive and not iterative logic. The purpose of the question is to aid learning recursion and not test you.

For Best Explanation : https://youtu.be/EohIyWnQYBY?list=PL-Jc9J83PIiFxaBahjslhBD1LiJAV7nKs

Input Format
A number x
A number n
Output Format
x raised to the power n

Constraints
1 <= x <= 10
0 <= n <= 9

Sample Input
2
5
Sample Output
32