Update and rename _218-Minimum In An Array to _157-Minimum In An Array
Given an array A[] of size n. The task is to find the smallest element in it.

Example 1:

Input:
n = 5
A[] = {1, 8, 7, 56, 90}
Output:
1
Explanation:
The smallest element of given array is 1.
 

Example 2:

Input:
n = 7
A[] = {1, 2, 0, 3, 2, 4, 5}
Output:
0
Explanation:
The smallest element of given array is 0.
 

Your Task:  
You don't need to read input or print anything. Your task is to complete the function smallest() which takes the array A[] and its size n as inputs and returns the minimum element in the array.