Update _233-Count Pairs In String
We'll say that a "pair" in a string is two instances of a char separated by a char. So "AxA" the A's make a pair. Pair's can overlap, so "AxAxA" contains 3 pairs -- 2 for A and 1 for x. 
Recursively compute the number of pairs in the given string.

(https://codingbat.com/prob/p154048)

countPairs("axa") → 1
countPairs("axax") → 2
countPairs("axbx") → 1
countPairs("ihjxhh") → 0
countPairs("ihihhh") → 3
countPairs("hi") → 0
countPairs("aaa") → 1