Update _182-Alternate Signed Numbers In Array
You are given an array arr. For every element in the array, the elements adjacent to it should be of opposite sign. Check if is true or not.

Note Consider 0 as a positive signed number.

Input Format
The first line contains an integer N, which represents the size of the array.

The second line contains N space-separated integers representing the array arr

Output Format
Print Yes if the array contains alternate signed numbers else, print No.

Example 1
Input

5
1 -2 3 -4 5
Output

Yes
Explanation

We have N = 5.

The digits are positive, negative, positive, negative, and positive, which is alternate

Example 2
Input

3
1 -1 -1
Output

No
Explanation

We have N = 3.

The digits are positive, negative, and negative, which is not alternate

Constraints
1 <= N <= 100000

-100000 <= arr[i] <= 100000